Return mocked data, even if the contract does not exist
