+++ crates/cheatnet/src/execution/cairo1_execution.rs
@@ -1,5 +1,3 @@
-use std::collections::HashSet;
-
 use super::syscalls::CheatableSyscallHandler;
 use crate::state::CheatcodeState;
 use blockifier::{
@@ -10,15 +8,14 @@ use blockifier::{
         },
         contract_class::{ContractClassV1, EntryPointV1},
         entry_point::{
-            CallEntryPoint, CallExecution, CallInfo, EntryPointExecutionContext,
-            EntryPointExecutionResult, ExecutionResources, Retdata,
+            CallEntryPoint, CallInfo, EntryPointExecutionContext, EntryPointExecutionResult,
+            ExecutionResources,
         },
         errors::{EntryPointExecutionError, VirtualMachineExecutionError},
         execution_utils::Args,
     },
     state::state_api::State,
 };
-use cairo_vm::vm::runners::cairo_runner::ExecutionResources as VmExecutionResources;
 use cairo_vm::{
     hint_processor::hint_processor_definition::HintProcessor,
     vm::{
@@ -26,7 +23,6 @@ use cairo_vm::{
         vm_core::VirtualMachine,
     },
 };
-use starknet_api::hash::StarkFelt;
 
 // blockifier/src/execution/cairo1_execution.rs:48 (execute_entry_point_call)
 pub fn execute_entry_point_call_cairo1(
@@ -37,25 +33,6 @@ pub fn execute_entry_point_call_cairo1(
     resources: &mut ExecutionResources,
     context: &mut EntryPointExecutionContext,
 ) -> EntryPointExecutionResult<CallInfo> {
-    // region: Modified blockifier code
-    if let Some(ret_data) = get_ret_data_by_call_entry_point(&call, cheatcode_state) {
-        return Ok(CallInfo {
-            call,
-            execution: CallExecution {
-                retdata: Retdata(ret_data.clone()),
-                events: vec![],
-                l2_to_l1_messages: vec![],
-                failed: false,
-                gas_consumed: 0,
-            },
-            vm_resources: VmExecutionResources::default(),
-            inner_calls: vec![],
-            storage_read_values: vec![],
-            accessed_storage_keys: HashSet::new(),
-        });
-    }
-    // endregion
-
     let VmExecutionContext {
         mut runner,
         mut vm,
@@ -136,19 +113,4 @@ pub fn cheatable_run_entry_point(
     )?;
 
     Ok(())
-}
-
-fn get_ret_data_by_call_entry_point<'a>(
-    call: &CallEntryPoint,
-    cheatcode_state: &'a CheatcodeState,
-) -> Option<&'a Vec<StarkFelt>> {
-    if let Some(contract_address) = call.code_address {
-        if let Some(contract_functions) = cheatcode_state.mocked_functions.get(&contract_address) {
-            let entrypoint_selector = call.entry_point_selector;
-
-            let ret_data = contract_functions.get(&entrypoint_selector);
-            return ret_data;
-        }
-    }
-    None
 }
