+++ crates/cheatnet/tests/cheatcodes/mock_call.rs
@@ -6,6 +6,7 @@ use crate::{
 use cairo_felt::Felt252;
 use cheatnet::rpc::call_contract;
 use conversions::StarknetConversions;
+use starknet_api::core::ContractAddress;
 
 #[test]
 fn mock_call_simple() {
@@ -355,6 +356,26 @@ fn mock_call_two_methods() {
     assert_success!(output, ret_data);
 
     let output = call_contract(&contract_address, &selector2, &[], &mut state).unwrap();
+
+    assert_success!(output, ret_data);
+}
+
+#[test]
+fn mock_call_nonexisting_contract() {
+    let mut state = create_cheatnet_state();
+
+    let selector = felt_selector_from_name("get_thing");
+    let ret_data = vec![Felt252::from(123)];
+
+    let contract_address = ContractAddress::from(218_u8);
+
+    state.start_mock_call(
+        contract_address,
+        &"get_thing".to_owned().to_felt252(),
+        &ret_data,
+    );
+
+    let output = call_contract(&contract_address, &selector, &[], &mut state).unwrap();
 
     assert_success!(output, ret_data);
 }
